+++ web/src/hooks/queries/useProjectViews.ts
@@ -1,0 +1,18 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { getProjectViews } from '../../api/project';
+import { useSession } from '../useSession';
+
+export const useProjectViews = (
+  namespace: string | undefined,
+  project: string | undefined,
+  tag: string | undefined,
+) => {
+  const session = useSession();
+  const query = useQuery({
+    queryKey: [namespace, project, tag, 'views'],
+    queryFn: () => getProjectViews(namespace!, project!, tag, session.jwt || ''),
+    enabled: !!namespace && !!project,
+  });
+  return query;
+};