+++ web/src/pages/Home.tsx
@@ -34,7 +34,12 @@ function Home() {
   const limit = 3;
   const { data: largestNamespaces } = useBiggestNamespace(limit);
 
-  const { data: exampleSamples } = useSampleTable('databio', 'example', 'default');
+  const { data: exampleSamples } = useSampleTable({
+    namespace: 'databio',
+    project: 'example',
+    tag: 'default',
+    enabled: true,
+  });
 
   return (
     <PageLayout>